+++ sdk/keyvault/azcertificates/CHANGELOG.md
@@ -1,4 +1,4 @@
 # Release History
 
 ## 0.1.0 (Unreleased)
-* This is the initial release of the `azkeys` library
+* This is the initial release of the `azcertificates` library